requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.11.1
Markdown==3.3.6
mccabe==0.6.1
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
psycopg2-binary>=2.8
gunicorn==20.1.0
django-extensions
pytest-django